# evmabi-backfill: Optimize db queries