--- a/.changelog/935.bugfix.md
+++ b/.changelog/935.bugfix.md
@@ -1,0 +1,1 @@
+evmabi-backfill: Optimize db queries